data-processor/requirements.txt
# Core dependencies with looser versioning
sentence-transformers>=2.2.0
langchain>=0.0.267
langchain-text-splitters>=0.0.1
pdfminer.six>=20221105
qdrant-client>=1.5.0

# Required dependencies for text processing
# Reduce dependencies to minimize conflicts
beautifulsoup4>=4.11.1
lxml>=4.9.2
markdown>=3.4.3
python-magic>=0.4.27

# Simplified HTML parsing for documents
html2text>=2020.1.16

# Use a simpler approach instead of unstructured
# unstructured has many dependencies that often conflict
pypdf>=3.15.1
chardet>=5.1.0
